chore: tableService package변경
